+++ android/src/main/kotlin/io/adrop/adrop_ads/bridge/AdropMethod.kt
@@ -16,6 +16,7 @@ object AdropMethod {
     const val LOG_EVENT = "logEvent"
     const val SEND_EVENT = "sendEvent"
     const val PERFORM_CLICK = "performClick"
+    const val REGISTER_WEB_VIEW = "registerWebView"
 
     // Consent methods
     const val REQUEST_CONSENT_INFO_UPDATE = "requestConsentInfoUpdate"